Rename localRPCs to RPCs
