+++ api/src/index.ts
@@ -1,5 +1,5 @@
 export * from "./api";
 export * from "./component-id";
-export * as localRPCs from "./local";
 export * from "./persisted-state";
 export * from "./rpc";
+export * as RPCs from "./rpcs";
