+++ app-foundations/src/components/UpdateBanner/UpdateBannerState.ts
@@ -1,4 +1,4 @@
-import type { localRPCs, ResponseOf } from "@previewjs/api";
+import type { ResponseOf, RPCs } from "@previewjs/api";
 import axios from "axios";
 import { makeAutoObservable, runInAction } from "mobx";
 import type { PersistedStateController } from "../../state/PersistedStateController";
@@ -13,7 +13,7 @@ export class UpdateBannerState {
     makeAutoObservable(this);
   }
 
-  async start(appInfo: ResponseOf<typeof localRPCs.GetInfo>["appInfo"]) {
+  async start(appInfo: ResponseOf<typeof RPCs.GetInfo>["appInfo"]) {
     try {
       const request: CheckVersionRequest = {
         appInfo,
